ct_cleanup:added option for high and/or low Cond Range

+++ R/CT_cleanup.R
@@ -11,9 +11,10 @@
 #' @param tf.write Logical parameter indicating whether to save output files in an output folder. Default = FALSE.
 #' @param tf.recursive Logical parameter indicating whether to search within folders at the data.path. Default = FALSE
 #' @param hobo.cal Logical parameter indicating whether the .hobo file has been pre-calibrated using HOBOware software and exported as a processed .csv. Default = FALSE
+#' @param cond.range Character string or object indicating whether conductivity was recorded in High ("high"), Low ("low"), or both ("highlow") ranges. Default = "high"
 #' @return For every imported CT data file, one tidied file with temperature-compensated conductance is exported and returned
 #' @export
-CT_cleanup<-function(data.path, output.path, path.pattern, tf.write = FALSE, tf.recursive = FALSE, hobo.cal = FALSE){
+CT_cleanup<-function(data.path, output.path, path.pattern, tf.write = FALSE, tf.recursive = FALSE, hobo.cal = FALSE, cond.range = "high"){
 
   if(hobo.cal == FALSE){
   # Create a list of all files within the directory folder
@@ -39,6 +40,7 @@ CT_cleanup<-function(data.path, output.path, path.pattern, tf.write = FALSE, tf.
     condCal <- file.names %>%
       purrr::map_dfr(~ readr::read_csv(file.path(data.path, .), skip=1, col_names=T)) # read all csv files at the file path, skipping 1 line of metadata and bind together
 
+    if(cond.range == "high"){
     condCal<-condCal %>%
       dplyr::select(contains('Date'), contains("High Range"), contains("Temp")) %>%
       dplyr::mutate(List.ID=Data_ID) %>%  # add column for file ID
@@ -48,21 +50,46 @@ CT_cleanup<-function(data.path, output.path, path.pattern, tf.write = FALSE, tf.
       tidyr::drop_na() %>%
       tidyr::separate(col = 'List.ID', into = c('List.ID',NA), sep = ".csv", remove = T) %>%  # remove the '.csv'
       dplyr::mutate(date = lubridate::mdy_hms(date))
-
+    } else if(cond.range == "low") {
+      condCal<-condCal %>%
+        dplyr::select(contains('Date'), contains("Low Range"), contains("Temp")) %>%
+        dplyr::mutate(List.ID=Data_ID) %>%  # add column for file ID
+        dplyr::rename(date=contains("Date"),
+                      TempInSitu=contains("Temp"),
+                      E_Conductivity=contains("Low Range")) %>%
+        tidyr::drop_na() %>%
+        tidyr::separate(col = 'List.ID', into = c('List.ID',NA), sep = ".csv", remove = T) %>%  # remove the '.csv'
+        dplyr::mutate(date = lubridate::mdy_hms(date))
+    } else if(cond.range == "highlow") {
+      condCal<-condCal %>%
+        dplyr::select(contains('Date'), contains("High Range"), contains("Low Range"), contains("Temp")) %>%
+        dplyr::mutate(List.ID=Data_ID) %>%  # add column for file ID
+        dplyr::rename(date=contains("Date"),
+                      TempInSitu=contains("Temp"),
+                      E_Conductivity_High=contains("High Range"),
+                      E_Conductivity_Low=contains("Low Range")) %>%
+        tidyr::drop_na() %>%
+        tidyr::separate(col = 'List.ID', into = c('List.ID',NA), sep = ".csv", remove = T) %>%  # remove the '.csv'
+        dplyr::mutate(date = lubridate::mdy_hms(date))
+    }
 
     full_df <- full_df %>%
       dplyr::full_join(condCal) # save your dataframes into a larger df
 
     # conditional write.csv at output path
     if(tf.write == TRUE) {
+      if(exists(output.path == T)){
       write.csv(condCal, paste0(output.path,'/',Data_ID,'_tidy.csv'))
+      } else {
+      write.csv(condCal, paste0(data.path,'/',Data_ID,'_tidy.csv'))
+      }
     }
    }
 
-
   full_df <- full_df %>%
     dplyr::rename(LoggerID = List.ID)
-  } else { # end of if(hobo.cal == FALSE)
+} else { # end of if(hobo.cal == FALSE)
+
     # Create a list of all files within the directory folder
     file.names.Cal<-basename(list.files(data.path, pattern = path.pattern, recursive = tf.recursive)) #list all csv file names in the folder and subfolders
 
@@ -86,7 +113,7 @@ CT_cleanup<-function(data.path, output.path, path.pattern, tf.write = FALSE, tf.
 
       condCal <- file.names %>%
         purrr::map_dfr(~ readr::read_csv(file.path(data.path, .), skip=1, col_names=T)) # read all csv files at the file path, skipping 1 line of metadata and bind together
-
+    if(cond.range == "high") {
       condCal<-condCal %>%
         dplyr::select(contains('Date'), contains("High Range"), contains("Temp"), contains("Salinity")) %>%
         dplyr::mutate(List.ID=Data_ID) %>%  # add column for file ID
@@ -97,21 +124,48 @@ CT_cleanup<-function(data.path, output.path, path.pattern, tf.write = FALSE, tf.
         tidyr::drop_na() %>%
         tidyr::separate(col = 'List.ID', into = c('List.ID',NA), sep = ".csv", remove = T) %>%  # remove the '.csv'
         dplyr::mutate(date = lubridate::mdy_hms(date))
-
+    } else if(cond.range == "low") {
+      condCal<-condCal %>%
+        dplyr::select(contains('Date'), contains("Low Range"), contains("Temp"), contains("Salinity")) %>%
+        dplyr::mutate(List.ID=Data_ID) %>%  # add column for file ID
+        dplyr::rename(date=contains("Date"),
+                      TempInSitu=contains("Temp"),
+                      E_Conductivity=contains("Low Range"),
+                      Salinity_psu=contains("Salinity")) %>%
+        tidyr::drop_na() %>%
+        tidyr::separate(col = 'List.ID', into = c('List.ID',NA), sep = ".csv", remove = T) %>%  # remove the '.csv'
+        dplyr::mutate(date = lubridate::mdy_hms(date))
+    } else if(cond.range == "highlow") {
+      condCal<-condCal %>%
+        dplyr::select(contains('Date'),contains("Low Range"), contains("High Range"), contains("Temp"), contains("Salinity")) %>%
+        dplyr::mutate(List.ID=Data_ID) %>%  # add column for file ID
+        dplyr::rename(date=contains("Date"),
+                      TempInSitu=contains("Temp"),
+                      E_Conductivity_High=contains("High Range"),
+                      E_Conductivity_Low=contains("Low Range"),
+                      Salinity_psu=contains("Salinity")) %>%
+        tidyr::drop_na() %>%
+        tidyr::separate(col = 'List.ID', into = c('List.ID',NA), sep = ".csv", remove = T) %>%  # remove the '.csv'
+        dplyr::mutate(date = lubridate::mdy_hms(date))
+    }
 
       full_df <- full_df %>%
         dplyr::full_join(condCal) # save your dataframes into a larger df
 
       # conditional write.csv at output path
       if(tf.write == TRUE) {
-        write.csv(condCal, paste0(output.path,'/',Data_ID,'_tidy.csv'))
+        if(exists(output.path == T)){
+          write.csv(condCal, paste0(output.path,'/',Data_ID,'_tidy.csv'))
+        } else {
+          write.csv(condCal, paste0(data.path,'/',Data_ID,'_tidy.csv'))
+        }
       }
     }
 
 
     full_df <- full_df %>%
       dplyr::rename(LoggerID = List.ID)
-  } # end of if(hobo.csv == TRUE)
+} # end of if(hobo.csv == TRUE)
 
   return(full_df) # return full dataframe
 }
